Store proxy address as IP address in cache entry
This ensures the IP address fields of any crawl log and warc record
entry is an actual IP address.
